requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 飞书SDK
lark-oapi==1.5.3

# Claude SDK
claude-agent-sdk==0.1.39

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1

# 工具库
structlog==23.2.0
prometheus-client==0.19.0
python-dotenv==1.0.0

# 额外依赖
python-multipart==0.0.6
